added overloaded methods writeToSRAM()/readFromSRAM() for writing/reading several bytes in one go
Added methods readFromSRAM() and writeToSRAM() which make it possible to to read/write several bytes from/to SRAM in a single SPI transaction.

This is a bit faster and (arguably) more readable than looping over SRAM addresses in the main sketch.

writeToSRAM() takes a pointer to an array to get the values from and the number of values to write.

readFromSRAM() takes a pointer to an array where the values from SRAM are written to and the number of values.